pyproject.toml
[project]
name = "pycroner"
version = "0.1.5"
description = "A multi-platform cron-like job runner in Python"
authors = [{ name="Marin Luic", email="marin.luic@gmail.com" }]
license = "GPL-3.0"
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
  "pyyaml>=6.0"
]


classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Operating System :: OS Independent",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Topic :: System :: Systems Administration",
  "Topic :: Utilities"
]



[project.scripts]
pycroner = "pycroner.__main__:cli_entrypoint"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.urls]
Homepage = "https://github.com/AcidBurnHen/pycroner"
Repository = "https://github.com/AcidBurnHen/pycroner"
